# perf(compiler): Cache expensive computation during mono