That's probably it for this perf uplift, most of the low-hanging fruits have been picked.

## Before
| Stage | Median time (ms) | Min time (ms) | Max time (ms) | Average time (ms) | Standard deviation (ms) |
| --- | --: | --: | --: | --: | --: |
| Init | 0 | 0 | 0 | 0.00 | 0.00 |
| Parse | 122 | 117 | 153 | 122.16 | 3.74 |
| Pass1 | 8 | 7 | 12 | 8.11 | 0.50 |
| Ast | 64 | 62 | 76 | 64.72 | 1.78 |
| Mono | 87 | 85 | 117 | 87.43 | 2.93 |
| Optimizations | 4 | 4 | 6 | 4.28 | 0.47 |
| Codegen | 26 | 25 | 39 | 26.11 | 1.46 |
| **TOTAL** | 318 | 311 | 352 | 319.23 | 6.20 |

## After
| Stage | Median time (ms) | Min time (ms) | Max time (ms) | Average time (ms) | Standard deviation (ms) |
| --- | --: | --: | --: | --: | --: |
| Init | 0 | 0 | 0 | 0.00 | 0.00 |
| Parse | 104 | 101 | 137 | 104.87 | 4.50 |
| Pass1 | 8 | 7 | 16 | 8.18 | 0.83 |
| Ast | 61 | 60 | 114 | 62.34 | 5.39 |
| Mono | 63 | 61 | 97 | 63.77 | 3.74 |
| Optimizations | 4 | 4 | 9 | 4.18 | 0.51 |
| Codegen | 26 | 25 | 36 | 26.79 | 1.72 |
| **TOTAL** | 267 | 261 | 383 | 270.12 | 12.47 |